requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
tavily-python>=0.5.0
streamlit>=1.31.0
python-dotenv>=1.0.0
typing-extensions>=4.9.0
pydantic>=2.6.0
